ConfigurationFactory has extracted interface for own implementation
