+++ dropwizard-core/src/test/java/io/dropwizard/server/DefaultServerFactoryTest.java
@@ -4,7 +4,8 @@ import com.codahale.metrics.MetricRegistry;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.io.CharStreams;
 import com.google.common.io.Resources;
-import io.dropwizard.configuration.ConfigurationFactory;
+
+import io.dropwizard.configuration.YamlConfigurationFactory;
 import io.dropwizard.jackson.DiscoverableSubtypeResolver;
 import io.dropwizard.jackson.Jackson;
 import io.dropwizard.jersey.errors.EarlyEofExceptionMapper;
@@ -61,9 +62,9 @@ public class DefaultServerFactoryTest {
                                                            SyslogAppenderFactory.class,
                                                            HttpConnectorFactory.class);
 
-        http = new ConfigurationFactory<>(DefaultServerFactory.class,
-                                          BaseValidator.newValidator(),
-                                          objectMapper, "dw")
+        http = new YamlConfigurationFactory<>(DefaultServerFactory.class,
+                                              BaseValidator.newValidator(),
+                                              objectMapper, "dw")
                 .build(new File(Resources.getResource("yaml/server.yml").toURI()));
     }
