+++ dropwizard-db/src/test/java/io/dropwizard/db/DataSourceConfigurationTest.java
@@ -1,7 +1,8 @@
 package io.dropwizard.db;
 
 import com.google.common.io.Resources;
-import io.dropwizard.configuration.ConfigurationFactory;
+
+import io.dropwizard.configuration.YamlConfigurationFactory;
 import io.dropwizard.jackson.Jackson;
 import io.dropwizard.jersey.validation.Validators;
 import io.dropwizard.util.Duration;
@@ -108,7 +109,7 @@ public class DataSourceConfigurationTest {
     }
 
     private DataSourceFactory getDataSourceFactory(String resourceName) throws Exception {
-        return new ConfigurationFactory<>(DataSourceFactory.class,
+        return new YamlConfigurationFactory<>(DataSourceFactory.class,
                 Validators.newValidator(), Jackson.newObjectMapper(), "dw")
                 .build(new File(Resources.getResource(resourceName).toURI()));
     }
