+++ dropwizard-metrics-graphite/src/test/java/io/dropwizard/metrics/graphite/GraphiteReporterFactoryTest.java
@@ -4,7 +4,8 @@ import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.graphite.Graphite;
 import com.codahale.metrics.graphite.GraphiteReporter;
 import com.codahale.metrics.graphite.GraphiteUDP;
-import io.dropwizard.configuration.ConfigurationFactory;
+
+import io.dropwizard.configuration.YamlConfigurationFactory;
 import io.dropwizard.jackson.DiscoverableSubtypeResolver;
 import io.dropwizard.jackson.Jackson;
 import io.dropwizard.validation.BaseValidator;
@@ -37,8 +38,8 @@ public class GraphiteReporterFactoryTest {
 
     @Test
     public void createDefaultFactory() throws Exception {
-        final GraphiteReporterFactory factory = new ConfigurationFactory<>(GraphiteReporterFactory.class,
-            BaseValidator.newValidator(), Jackson.newObjectMapper(), "dw")
+        final GraphiteReporterFactory factory = new YamlConfigurationFactory<>(GraphiteReporterFactory.class,
+             BaseValidator.newValidator(), Jackson.newObjectMapper(), "dw")
             .build();
         assertThat(factory.getFrequency()).isEqualTo(Optional.empty());
     }
