+++ dropwizard-metrics/src/test/java/io/dropwizard/metrics/MetricsFactoryTest.java
@@ -2,7 +2,8 @@ package io.dropwizard.metrics;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.io.Resources;
-import io.dropwizard.configuration.ConfigurationFactory;
+
+import io.dropwizard.configuration.YamlConfigurationFactory;
 import io.dropwizard.jackson.Jackson;
 import io.dropwizard.logging.BootstrapLogging;
 import io.dropwizard.util.Duration;
@@ -20,10 +21,10 @@ public class MetricsFactoryTest {
     }
 
     private final ObjectMapper objectMapper = Jackson.newObjectMapper();
-    private final ConfigurationFactory<MetricsFactory> factory =
-            new ConfigurationFactory<>(MetricsFactory.class,
-                                       BaseValidator.newValidator(),
-                                       objectMapper, "dw");
+    private final YamlConfigurationFactory<MetricsFactory> factory =
+            new YamlConfigurationFactory<>(MetricsFactory.class,
+                                           BaseValidator.newValidator(),
+                                           objectMapper, "dw");
     private MetricsFactory config;
 
     @Before
